+++ src/VertexBufferLayout.cpp
@@ -3,13 +3,13 @@
 using namespace giygas;
 using namespace std;
 
-void VertexBufferLayout::set_channel(int index, int size) {
+void VertexBufferLayout::set_channel(size_t index, size_t size) {
     if (index >= _channels.size()) {
         _channels.resize(index + 1);
     }
     _channels[index] = size;
 }
 
-const vector<int>& VertexBufferLayout::get_channels() const {
+const vector<size_t>& VertexBufferLayout::get_channels() const {
     return _channels;
 }
